add context to pipe constructor
Context now is required to build a pipe. It allows to ensure that all pipe elements will be bound to the same context and using same buffer size.

The Routing type is renamed to Line. It's still used to only define a sequence of DSP components.

The Line type is renamed to Pipe. Now it defines a chain of DSP component sequences bound together (multiple Routes from the old notation).

The Pipe type is renamed to Runner. It represents a pipe in a running state.